# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Solana Concepts Sandbox API",
    "version": "1.0.0",
    "description": "API documentation for dHealth&lt;&gt;Solana Sandbox concepts",
    "contact": {
      "name": "dHealth Team",
      "url": "https://dhealth.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://unrevised-framing-silo.ngrok-free.dev",
      "description": "Development server"
    },
    {
      "url": "https://dapps.evi.as",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
    },
    "securitySchemes": {
      "WalletAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "wallet",
        "description": "Solana wallet address for authorization"
      }
    }
  },
...
  "paths": {
...
    "/api/v1/hcpconsole/build-attestation-tx": {
      "post": {
        "tags": [
          "HCPConsole"
        ],
        "summary": "Create a patient journey with provider attested AI prompt.",
        "description": "Creates an AI prompt for a patient journey which is SAS attested by a provider.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "wallet",
                  "credentialId",
                  "caseRef",
                  "promptHash",
                  "promptText"
                ],
                "properties": {
                  "wallet": {
                    "type": "string",
                    "description": "Solana wallet address"
                  },
                  "credentialId": {
                    "type": "string",
                    "description": "Credential sasCredentialId"
                  },
                  "caseRef": {
                    "type": "string",
                    "description": "Reference of the case"
                  },
                  "promptHash": {
                    "type": "string",
                    "description": "SHA256 hash of the AI prompt"
                  },
                  "promptText": {
                    "type": "string",
                    "description": "AI prompt content"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prompt attestation created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "file": {
                      "type": "object",
                      "properties": {
                        "txSig": {
                          "type": "string"
                        },
                        "credentialPda": {
                          "type": "string"
                        },
                        "schemaPda": {
                          "type": "string"
                        },
                        "attestationPda": {
                          "type": "string"
                        },
                        "isTestMode": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "description": "Access denied - wallet not authorized for credential"
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/api/v1/hcpconsole/prompts": {
      "get": {
        "tags": [
          "HCPConsole"
        ],
        "summary": "Retrieve hcp_prompts entries by sas_credential_id.",
        "description": "Retrieve hcp_prompts entries by sas_credential_id.",
        "parameters": [
          {
            "in": "query",
            "name": "credential",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The SAS Credential ID associated with prompts."
          }
        ],
        "responses": {
          "200": {
            "description": "Prompts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prompts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "wallet_address": {
                            "type": "string"
                          },
                          "sas_credential_id": {
                            "type": "string"
                          },
                          "case_ref": {
                            "type": "string"
                          },
                          "prompt_hash": {
                            "type": "string"
                          },
                          "transaction_signature": {
                            "type": "string"
                          },
                          "lastread_at": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/api/v1/hcpconsole/prompt": {
      "get": {
        "tags": [
          "HCPConsole"
        ],
        "summary": "Retrieve hcp_prompts.prompt_cipher in plaintext by case_ref",
        "description": "Retrieves hcp_prompts.prompt_cipher in plaintext by case_ref.",
        "parameters": [
          {
            "in": "query",
            "name": "caseRef",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The patient journey case reference."
          },
          {
            "in": "query",
            "name": "promptHash",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The prompt hash."
          },
          {
            "in": "query",
            "name": "enableMeta",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Whether to include metadata."
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prompt": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/Error"
          },
          "500": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
...
  }
}